docs, reorganization
