+++ x/data/server/testsuite/suite_test.go
@@ -1,6 +1,7 @@
 package testsuite
 
 import (
+	"github.com/regen-network/regen-ledger/testutil/server/configurator"
 	"testing"
 
 	"github.com/stretchr/testify/suite"
@@ -8,7 +9,6 @@ import (
 	"github.com/cosmos/cosmos-sdk/testutil/testdata"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 
-	servertestutil "github.com/regen-network/regen-ledger/testutil/server"
 	dataserver "github.com/regen-network/regen-ledger/x/data/server"
 )
 
@@ -16,7 +16,7 @@ func TestSuite(t *testing.T) {
 	key := sdk.NewKVStoreKey("ecocredit")
 	_, _, addr := testdata.KeyTestPubAddr()
 
-	configuratorFixture := servertestutil.NewConfiguratorFixture(t, []sdk.StoreKey{key}, []sdk.AccAddress{addr})
+	configuratorFixture := configurator.NewConfiguratorFixture(t, []sdk.StoreKey{key}, []sdk.AccAddress{addr})
 	dataserver.RegisterServices(key, configuratorFixture)
 	s := NewIntegrationTestSuite(configuratorFixture)
